+++ test/tiles/bigbed_test.py
@@ -1,5 +1,5 @@
 import clodius.tiles.bigbed as hgbb
-import clodius.tiles.bigwig as hgbw
+import clodius.tiles.utils as hgut
 import os.path as op
 
 
@@ -87,7 +87,7 @@ def test_natsorted():
     ]
 
     for test in chromname_tests:
-        sorted_output = hgbw.natsorted(test["input"])
+        sorted_output = hgut.natsorted(test["input"])
         assert (
             sorted_output == test["expected"]
         ), "Sorted output was %s\nExpected: %s" % (sorted_output, test["expected"])
